refactored shim out
