+++ cPecanRealign.c
@@ -13,9 +13,10 @@
 #include "pairwiseAligner.h"
 #include "multipleAligner.h"
 #include "commonC.h"
-#include "inc/shim.h"
+//#include "inc/shim.h"
 #include "inc/pairwiseAligner.h"
 #include "inc/multipleAligner.h"
+#include "../sonLib/C/inc/bioioC.h"
 
 void usage() {
     fprintf(stderr, "cPecanRelign [options] seq1[fasta] seq2[fasta], version 0.2\n");
