+++ impl/multipleAligner.c
@@ -663,8 +663,11 @@ static int64_t addMultipleAlignedPairs(StateMachine *sM, int64_t sequence1, int6
      */
     SeqFrag *seqFrag1 = stList_get(seqFrags, sequence1);
     SeqFrag *seqFrag2 = stList_get(seqFrags, sequence2);
-    stList *alignedPairs = getAlignedPairs(sM, seqFrag1->seq, seqFrag2->seq, nucleotide, pairwiseAlignmentBandingParameters, // hardwired nucleotide sequence type for now
-            seqFrag1->leftEndId != seqFrag2->leftEndId, seqFrag1->rightEndId != seqFrag2->rightEndId);
+    stList *alignedPairs = getAlignedPairs(sM, seqFrag1->seq, seqFrag2->seq,
+                                           //nucleotide,
+                                           pairwiseAlignmentBandingParameters,
+                                           seqFrag1->leftEndId != seqFrag2->leftEndId,
+                                           seqFrag1->rightEndId != seqFrag2->rightEndId);
     alignedPairs = reweightAlignedPairs2(alignedPairs, seqFrag1->length, seqFrag2->length, pairwiseAlignmentBandingParameters->gapGamma);
     int64_t distance = getAlignmentScore(alignedPairs, seqFrag1->length, seqFrag2->length);
     convertAlignedPairsToMultipleAlignedPairs(alignedPairs, multipleAlignedPairs, sequence1, sequence2);
