+++ tests/shimTests.c
@@ -4,13 +4,14 @@
 #include <string.h>
 #include <stdlib.h>
 #include <stdio.h>
-#include "shim.h"
+//#include "shim.h"
 #include "CuTest.h"
 #include "randomSequences.h"
-#include "../inc/shim.h"
+//#include "../inc/shim.h"
 #include "../../sonLib/lib/CuTest.h"
 #include "../inc/stateMachine.h"
 #include "../inc/pairwiseAligner.h"
+#include "../inc/emissionMatrix.h"
 
 static void test_sequenceConstruct(CuTest* testCase) {
     char *tS = getRandomSequence(100);
@@ -46,7 +47,7 @@ static void test_kmers(CuTest* testCase) {
         CuAssertStrEquals(testCase, &testKmers[x], kmer);
     }
 }
-
+/*
 static void test_events(CuTest* testCase) {
     char testKmers[12][6] = { "GATAC", "ATACA", "TACAG", "ACAGA", "CAGAT", "AGATA",
                               "GATAC", "ATACA", "TACAG", "ACAGA", "CAGAT", "AGATA" };
@@ -63,15 +64,53 @@ static void test_events(CuTest* testCase) {
         CuAssertDblEquals(testCase, testMeans[x], ev->mean, 0.0);
     }
 }
+*/
 
+static void test_getKmerIndex(CuTest* testCase) {
+    char kmer2Set[25][3] = {"AA", "AC", "AG", "AT", "AN", "CA", "CC", "CG", "CT", "CN",
+                           "GA", "GC", "GG", "GT", "GN", "TA", "TC", "TG", "TT", "TN",
+                           "NA", "NC", "NG", "NT", "NN"};
+
+    for (int64_t i = 0; i < NUM_OF_KMERS; i++) {
+        int64_t kmer_index_result = getKmerIndex(kmer2Set[i]);
+        CuAssertIntEquals(testCase, kmer_index_result, i);
+    }
+}
 
 CuSuite* shimTestSuite() {
     CuSuite* suite = CuSuiteNew();
 
     SUITE_ADD_TEST(suite, test_sequenceConstruct);
+    SUITE_ADD_TEST(suite, test_getKmerIndex);
 //    SUITE_ADD_TEST(suite, test_chars);
 //    SUITE_ADD_TEST(suite, test_kmers);
 //    SUITE_ADD_TEST(suite, test_events);
 
     return suite;
 }
+
+
+/*
+ * static void test_getKmerIndex(CuTest* testCase) {
+    char* testA = "A";
+    int64_t iA = getKmerIndex(testA);
+
+    char* testC = "C";
+    int64_t iC = getKmerIndex(testC);
+
+    char* testG = "G";
+    int64_t iG = getKmerIndex(testG);
+
+    char* testT = "T";
+    int64_t iT = getKmerIndex(testT);
+
+    char* testN = "N";
+    int64_t iN = getKmerIndex(testN);
+
+    printf("%s, %lld\n", testA, iA);
+    printf("%s, %lld\n", testC, iC);
+    printf("%s, %lld\n", testG, iG);
+    printf("%s, %lld\n", testT, iT);
+    printf("%s, %lld\n", testN, iN);
+}
+ */
